requirements.txt
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
